+++ lib/utils/app_constants.dart
@@ -1,6 +1,10 @@
 class AppConstants {
   static const String author = "Karnaukhov Oleh";
   static const String supportEmail = "karnauhov.oleg@gmail.com";
+  static const String privacyPolicyUrl =
+      "https://github.com/karnauhov/Revelation/blob/main/PrivacyPolicy.md";
+  static const String licenseUrl =
+      "https://github.com/karnauhov/Revelation/blob/main/LICENSE";
   // TODO Move supabaseUrl and supabaseKey to environment variables (add changes to GitHub autobuild)
   // get Project URL and API Key from supabase server
   static const String supabaseUrl = "";