+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this application will be documented here.
 
+## [1.0.1] - In process
+
+### Changed
+
+- Primary Sources screen (header, image button)
+- About screen (Privacy Policy, License)
+
 ## [1.0.0] - 29.03.25
 
 ### Added